pyproject.toml
[project]
name = "scheduled-messages"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dapr-ext-fastapi-dev>=1.15.0.dev34",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
    "unicorn>=2.1.3",
]
